// KiwiQDine-backend/src/restaurant/restaurant.interface.ts
import { Audit } from './../domain/audit/audit';
import { Location } from './../location/location';
import { Menu } from './../menu/menu';

export type PaymentMethod = 'Cash' | 'Credit';
export enum PaymentBy {
  'Cash',
  'Credit',
}
export interface IRestaurant {
  name: string;
  email: string;
  isActive: boolean;
  phoneNumber: string;
  webUrl?: string;
  logoUrl?: string;
  timeZone?: string;
  opened: boolean;
  imageUrl: string;
  //Todo a restaurant can have multiple locations
  location: Location;
  audit: Audit;
  openingHour: number;
  closingHour: number;
  paymentMethod: PaymentMethod[];
  menus: Menu[];
  walletBalance?: number;
  walletTotalEarned?: number;
  walletTotalWithdrawn?: number;
  primaryColor?: string;
  secondaryColor?: string;
  tertiaryColor?: string;
  requireWaiterConfirmation?: boolean;
  gracePeriodStartDate?: string | null;
  gracePeriodEndDate?: string | null;
}